took a different tactic; rather than having an animal form child of animaledit and animalnew, I merged all three one form file
